move SumAggregationResult implementation into own file
