+++ searchlib/src/vespa/searchlib/aggregation/countaggregationresult.cpp
@@ -1,6 +1,6 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "aggregation.h"
+#include "countaggregationresult.h"
 
 #include <vespa/document/fieldvalue/document.h>
 #include <vespa/searchlib/expression/resultvector.h>
